+++ crates/containerd-shim-wasm/src/sys/unix/container/instance.rs
@@ -34,13 +34,12 @@ pub struct Instance<E: Engine> {
     _phantom: PhantomData<E>,
 }
 
-impl<E: Engine> SandboxInstance for Instance<E> {
+impl<E: Engine + Default> SandboxInstance for Instance<E> {
     type Engine = E;
 
     #[cfg_attr(feature = "tracing", tracing::instrument(parent = tracing::Span::current(), skip_all, level = "Info"))]
-    fn new(id: String, cfg: Option<&InstanceConfig<Self::Engine>>) -> Result<Self, SandboxError> {
-        let cfg = cfg.context("missing configuration")?;
-        let engine = cfg.get_engine();
+    fn new(id: String, cfg: &InstanceConfig) -> Result<Self, SandboxError> {
+        let engine = Self::Engine::default();
         let bundle = cfg.get_bundle().to_path_buf();
         let namespace = cfg.get_namespace();
         let rootdir = Path::new(DEFAULT_CONTAINER_ROOT_DIR).join(E::name());